Fix: move loader to a separate class
Here's why:
The loader was being created as an anonymous class inside of the MainActivity , and given that an Activity is going to be alive in its own lifecycle, there is a possibility that an instance of this class is still alive while the Activity's is not and then it can keep creating new instances of this loader and increasing the memory used, thus there is a probability of a memory leak.
To fix this, I moved this loader to a separate class,  this way this would live along with the Activity's lifecycle and prevent that potential memory leak.